Make ->eval not throw exceptions
RT #65348 made the eval method throw. After a bit of thought I think the
original approach of ->eval not generating exceptions but setting $@
makes the most sense. This means it can be treated in a similar way to
an eval { } block, rather than having to wrap the eval method in an eval
block.