Add persistent handle for functions and free magic
Previously V8 wasn't actually told we were holding on to the function,
so use a persistent handle. Add svt_magic to dispose of the function.

I'm not sure this is 100% correct still as the code is taking the value
out of the handle, I think we need to keep the handle around rather than
recreating it later...